Agent-Drone/requirements.txt
pyautogen
autogen
mavsdk
python-dotenv
